requirements.txt
coloredlogs==15.0.1
python-dotenv==0.19.2
pyyaml==6.0